fix review, extract locator_converter
